Unit tests

+++ Infrastructure/AdvisorRepository.cs
@@ -8,16 +8,17 @@ namespace AdvisorManagement.Infrastructure.Persistence
     public class AdvisorRepository : IAdvisorRepository
     {
         private readonly AppDbContext _context;
-        private readonly MRUCache<int, string> _cache;
+        private readonly MRUCache<int, Advisor> _cache;
 
-        public AdvisorRepository(AppDbContext context)
+        public AdvisorRepository(AppDbContext context, MRUCache<int, Advisor> cache)
         {
             _context = context;
-            MRUCache<int, string> cache = new MRUCache<int, string>(5);
+            _cache = cache;
         }
 
         public async Task AddAsync(Advisor advisor)
         {
+            advisor.HealthStatus = AdvisorRepository.GetColor();
             await _context.Advisors.AddAsync(advisor);
             await _context.SaveChangesAsync();
         }
@@ -26,17 +27,46 @@ namespace AdvisorManagement.Infrastructure.Persistence
         {
             _context.Advisors.Update(advisor);
             await _context.SaveChangesAsync();
+            _cache.Put(advisor.Id, advisor); // Update cache
         }
 
-        public async Task DeleteAsync(Advisor advisor)
+        public async Task DeleteAsync(int advisorId)
         {
-            _context.Advisors.Remove(advisor);
-            await _context.SaveChangesAsync();
+            var advisor = await _context.Advisors.FindAsync(advisorId);
+            if (advisor != null)
+            {
+                _context.Advisors.Remove(advisor);
+                await _context.SaveChangesAsync();
+                _cache.Delete(advisorId); // Remove from cache
+            };
         }
 
-        public Advisor GetAsync(int advisorId)
+        public async Task<Advisor> GetAsync(int advisorId)
         {
-            return _context.Advisors.FirstOrDefault(x => x.Id == advisorId);
+            // Check if advisor exists in cache
+            var advisor = _cache.TryGet(advisorId);
+            
+            if (advisor != null)
+                return advisor;
+
+            // If not found in cache, retrieve from database and add to cache
+            advisor = await _context.Advisors.FindAsync(advisorId);
+            if (advisor != null)
+                _cache.Put(advisorId, advisor);
+
+            return advisor;
+        }
+
+        public static string GetColor()
+        {
+            var hColors = new Dictionary<int, string>() { { 60, "Green" }, { 80, "Yellow" }, { 100, "Green" } };
+
+            var random = new Random();
+            int rndprob = random.Next(100);
+
+            var retColor = hColors.Last(x => rndprob <= x.Key);
+
+            return retColor.Value;
         }
     }
 
@@ -44,7 +74,8 @@ namespace AdvisorManagement.Infrastructure.Persistence
     {
         Task AddAsync(Advisor advisor);
         Task UpdateAsync(Advisor advisor);
-        Task DeleteAsync(Advisor advisor);
-        Advisor GetAsync(int advisorId);        
+        Task DeleteAsync(int advisorId);
+        Task<Advisor> GetAsync(int advisorId);
+
     }
 }
